added while loop for building meal plan, and set() for building grocery list
used a while loop to ensure the meal plan had the right number of
entries even if duplicate random selections were excluded, and a for
loop to check a meal wasn’t already in the meal plan before appending.
set() and some print formatting now also prints an attractive list with
no duplicates.